ref: Move CORS middleware
