+++ relay-server/src/endpoints/common.rs
@@ -1,15 +1,12 @@
 //! Common facilities for ingesting events through store-like endpoints.
 
-use std::time::Duration;
-
-use axum::http::{header, HeaderName, Method, StatusCode};
+use axum::http::{header, StatusCode};
 use axum::response::IntoResponse;
 use relay_general::protocol::{EventId, EventType};
 use relay_log::LogError;
 use relay_quotas::RateLimits;
 use relay_statsd::metric;
 use serde::Deserialize;
-use tower_http::cors::CorsLayer;
 
 use crate::actors::outcome::{DiscardReason, Outcome};
 use crate::actors::processor::{EnvelopeProcessor, ProcessMetrics};
@@ -303,36 +300,6 @@ pub fn event_id_from_items(items: &Items) -> Result<Option<EventId>, BadStoreReq
     }
 
     Ok(None)
-}
-
-/// Creates a preconfigured CORS middleware builder for store requests.
-///
-/// To configure CORS, register endpoints using `resource()` and finalize by calling `register()`, which
-/// returns an App. This configures POST as allowed method, allows default sentry headers, and
-/// exposes the return headers.
-pub fn cors() -> CorsLayer {
-    CorsLayer::new()
-        .allow_methods(Method::POST)
-        .allow_headers([
-            HeaderName::from_static("x-sentry-auth"),
-            HeaderName::from_static("x-requested-with"),
-            HeaderName::from_static("x-forwarded-for"),
-            HeaderName::from_static("origin"),
-            HeaderName::from_static("referer"),
-            HeaderName::from_static("accept"),
-            HeaderName::from_static("content-type"),
-            HeaderName::from_static("authentication"),
-            HeaderName::from_static("authorization"),
-            HeaderName::from_static("content-encoding"),
-            HeaderName::from_static("transfer-encoding"),
-        ])
-        .allow_origin(tower_http::cors::Any)
-        .expose_headers([
-            HeaderName::from_static("x-sentry-error"),
-            HeaderName::from_static("x-sentry-rate-limits"),
-            HeaderName::from_static("retry-after"),
-        ])
-        .max_age(Duration::from_secs(3600))
 }
 
 /// Queues an envelope for processing.
