+++ relay-server/src/middlewares/mod.rs
@@ -1,7 +1,9 @@
+mod cors;
 mod decompression;
 mod metrics;
 mod normalize_path;
 
+pub use self::cors::*;
 pub use self::decompression::*;
 pub use self::metrics::*;
 pub use self::normalize_path::*;
